requirements.txt
python-telegram-bot==20.7
huggingface_hub
python-dotenv
